Fix permission errors when building docker container
Also add `need-app=true` to `uwsgi.ini` for easier debugging when things
don't work. Because of the permission errors inside the container,
`uwsgi` was unable to load `bookmarks.wsgi` but it continued starting up
anyways. Adding `need-app=true` causes `uwsgi` to fail and exit if it
can't load the WSGI app for some reason. This makes it much easier to
catch startup errors.